+++ CHANGELOG.md
@@ -5,6 +5,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.25.0] - Not released
+### Added
+- New API for pinning and unpinning posts in groups and author's profile. The
+  new methods are:
+  - `POST /vN/posts/:postId/pin`
+  - `POST /vN/posts/:postId/unpin`
+   
+  In both methods the request body is `{ "target": "username" }`. The _target_
+  is the username of the target user or group, to which feed the post will be
+  pinned. User can pin only their own posts and only to their own feed or
+  managed groups. The _target_ may be omitted to pin the post to the user's
+  profile.
+
+  Serialized posts have a new `pinnedIn` field, if the post is pinned to some
+  feed. This field is an array of `{"targetId": UUID, "pinnedAt": ISODateTime}`
+  objects.
+
+- Pinned posts affect visibility of protected/private user feeds. These feeds
+  are not visible to anonymous/not-subscribed users, even if they have visible
+  posts. It is a special behavior of such feeds.
+  
+  For example: private user creates a post, which belongs to their feed and to
+  some public group. This post has public visibility, so it is visible to
+  anonymous users (in group, in home feed, or by direct link). But when
+  anonymous user requests the author's feed, they will not see any post.
+
+  Pinned posts break this behavior: they are visible to anyone, who can see
+  them. So private user can make some public posts, pin them to their feed, and
+  these posts will be visible on their page for any guest.
+
 ## [2.24.3] - 2025-08-14
 ### Fixed
 - Fix attachment sanitization when the file isn't found on S3 storage.